% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_network.R
\name{calculate_streamorder}
\alias{calculate_streamorder}
\title{Get Streamorder}
\usage{
calculate_streamorder(fl)
}
\arguments{
\item{flowline}{data.frame with dendritic ID and toID columns.}
}
\value{
data.frame with ID and order columns. Note the output will
likelt NOT be in the same sort order as in the input.
}
\description{
Applies a topological sort and calculates strahler stream order.
Algorithm: If more than one upstream flowpath has an order equal to the
maximum upstream order then the downstream flowpath is assigned the maximum
upstream order plus one. Otherwise it is assigned the max upstream order.
}
\examples{
source(system.file("extdata", "walker_data.R", package = "nhdplusTools"))

test_flowline <- prepare_nhdplus(walker_flowline, 0, 0, FALSE)

test_flowline <- data.frame(
  ID = test_flowline$COMID,
  toID = test_flowline$toCOMID)

(order <- get_streamorder(test_flowline))

walker_flowline <- left_join(walker_flowline, order, by = c("COMID" = "ID"))

plot(sf::st_geometry(walker_flowline), lwd = walker_flowline$order, col = "blue")

}
