% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplusTools.R
\docType{data}
\name{catchment_sample}
\alias{catchment_sample}
\title{Sample Catchments Walker River}
\format{An sf data.frame}
\source{
\url{https://www.epa.gov/waterdata/nhdplus-national-data}
}
\usage{
catchment_sample
}
\description{
Sample Catchments
}
\keyword{datasets}
