% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_flowlines.R
\name{collapse_flowlines}
\alias{collapse_flowlines}
\title{Collapse NHDPlus Network}
\usage{
collapse_flowlines(flines, thresh)
}
\arguments{
\item{flines}{data.frame with COMID, toCOMID, LENGTHKM, and TotDASqKM columns}

\item{thresh}{numeric a length threshold (km). Flowlines shorter than this will be eliminated}
}
\value{
A refactored network with merged up and down flowlines.
}
\description{
Refactors the NHDPlus flowline network, eliminating short and non-cconfluence flowlines.
}
