% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_headwaters.R
\name{collapse_headwaters}
\alias{collapse_headwaters}
\title{Collapse Short Headwaters}
\usage{
collapse_headwaters(flines, thresh, original_fline_atts)
}
\arguments{
\item{flines}{data.frame with COMID, toCOMID, LENGTHKM, and TotDASqKM columns}

\item{thresh}{numeric a length threshold (km). Flowlines shorter than this will be eliminated}
}
\value{
flines data.frame with short headwaters merged downsrteam
}
\description{
Refactors the NHDPlus flowline network, eliminating short and non-cconfluence flowlines.
}
