% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_outlets}
\alias{collapse_outlets}
\title{Collapse Outlets}
\usage{
collapse_outlets(flines, thresh, original_fline_atts)
}
\arguments{
\item{flines}{data.frame with COMID, toCOMID, LENGTHKM, and TotDASqKM columns}

\item{thresh}{numeric a length threshold (km). Flowlines shorter than this will be eliminated}

\item{original_fline_atts}{data.frame containing original fline attributes.}
}
\value{
flines data.frame with outlets collapsed to the given threshold.
}
\description{
Collapses outlet flowlines upstream. Returns an flines dataframe.
}
