% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_watershed_delineation_nhdplushr.R
\name{delineate_watersheds}
\alias{delineate_watersheds}
\title{delineate_watersheds}
\usage{
delineate_watersheds(points, nhdplusdir, crs = "epsg:4326")
}
\arguments{
\item{points}{an sf object of type POINT or MULTIPOINT containing the the locations
of desired catchment outlets}

\item{nhdplusdir}{directory location of nhdplushr data downloads}

\item{crs}{the coordinate reference system for geoprocessing. Defaults to epsg:4326}
}
\value{
watershed areas for each point
}
\description{
Wrapper function to find the catchments for all the points. Can take
points from multiple HUC4s
}
