% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geoconnex.R
\name{discover_geoconnex_reference}
\alias{discover_geoconnex_reference}
\title{discover geoconnex reference feature layers}
\usage{
discover_geoconnex_reference()
}
\value{
data.frame containing layers available and fields that are available to query.
}
\description{
Queries the geoconnex.us reference feature server for available layers and
attributes.
}
