% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nhd.R
\name{discover_huc12}
\alias{discover_huc12}
\title{Find WBD HUC 12 unit subsets}
\usage{
discover_huc12(AOI = NULL, id = NULL)
}
\arguments{
\item{AOI}{an 'area of interest' object (sf POINT or POLYGON)}

\item{id}{Search for WBD HUC12 features by ID}
}
\value{
simple feature polygon object
}
\description{
Subsets the WBD level 12 features by location (POINT), area (POLYGON), or set of IDs.
}
