% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_hydro.R
\name{discover_huc8}
\alias{discover_huc8}
\title{Find WBD HUC 08 unit subsets}
\usage{
discover_huc8(AOI = NULL, id = NULL)
}
\arguments{
\item{AOI}{sf (MULTI)POINT or (MULTI)POLYGON. An 'area of interest' can
be given as either a location (sf POINT) or area (sf POLYGON)
in any coordinate reference system.}

\item{id}{WBD HUC08 ID(s)}
}
\value{
simple feature polygon object
}
\description{
Subsets the WBD level 08 features by location (POINT),
area (POLYGON), or set of IDs.
}
