% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nldi.R
\name{discover_nldi_characteristics}
\alias{discover_nldi_characteristics}
\title{Discover Characteristics Metadata}
\usage{
discover_nldi_characteristics(type = "all")
}
\arguments{
\item{type}{character "all", "local", "total", or "divergence_routed".}
}
\description{
Provides access to metadata for characteristics that are returned by `get_nldi_characteristics()`.
}
\examples{
chars <- discover_nldi_characteristics()
names(chars)
head(chars$local, 10)
}
