% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nldi.R
\name{discover_nldi_sources}
\alias{discover_nldi_sources}
\title{Discover NLDI Sources}
\usage{
discover_nldi_sources(tier = "prod")
}
\arguments{
\item{tier}{character optional "prod" or "test"}
}
\value{
data.frame with three columns "source", "sourceName"
and "features"
}
\description{
Function to retrieve available feature
and data sources from the Network Linked Data Index.
}
\examples{
\donttest{
discover_nldi_sources()
}
}
