% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainstems.R
\name{download_mainstem_lookup}
\alias{download_mainstem_lookup}
\title{Download mainstem lookup table from ref_rivers}
\usage{
download_mainstem_lookup(
  type,
  path = get_mainstem_lookup_path(type),
  force = FALSE
)
}
\arguments{
\item{type}{character one of "nhdpv2" or "nhdphr"}

\item{path}{character path where the file should be saved. Default is a
persistent system data dir as retrieved by \link{hydrogeofetch_data_dir}.
Also see: \link{get_mainstem_lookup_path}}

\item{force}{logical. Force data re-download. Default = FALSE}
}
\value{
character path to cached data
}
\description{
Downloads and caches a geoconnex mainstem identifier lookup
table on your computer.
}
\details{
Source data comes from the ref_rivers GitHub release
\href{https://github.com/internetofwater/ref_rivers/releases}{here}. The
csv source data is downloaded once, converted to parquet, and cached in
the user data dir indicated by \link{hydrogeofetch_data_dir}.
}
\keyword{internal}
