% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhdplus.R
\name{download_nhdhr}
\alias{download_nhdhr}
\title{download NHDPlus HiRes}
\usage{
download_nhdhr(nhd_dir, hu_list, download_files = TRUE)
}
\arguments{
\item{nhd_dir}{character directory to save output into}

\item{hu_list}{character vector of hydrologic region(s) to download
can be hu02s and/or hu04s}
}
\value{
Paths to geodatabases created.
}
\description{
download NHDPlus HiRes
}
\examples{
download_nhdhr("", c("01", "0203"), download_files = FALSE)
}
