% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{downloader}
\alias{downloader}
\title{Downloader: Downloading National-Scale Hydro feature archives}
\usage{
downloader(outdir, type = "rf1")
}
\arguments{
\item{outdir}{The folder path where data should be downloaded and extracted}

\item{type}{The dataset to download and extract. Options include (1) nhdplus (2) wbd (3) rf1}
}
\value{
the path to the respective geodatabase or e00 file
}
\description{
This function downloads and decompresses staged seamless data representing various hydro features in CONUS.
Available options include the seamless NHDPlusV21, the Watershed Boundary dataset and the EPA Reach File (RF). The WBD and RF can be downloaded with no system dependencies.
To download and extract the NHDPlus users need to meet the following requirements: p7zip (MacOS), 7zip (windows)
}
\examples{
\dontrun{
outdir = "./data"
  downloader(outdir, "nhdplus")
  downloader(outdir, "wbd")
  downloader(outdir, "rf1")
}
}
