% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplusTools.R
\docType{data}
\name{fac_sample}
\alias{fac_sample}
\title{Sample Flow Accumulation Raster}
\format{A raster object}
\source{
\url{https://www.epa.gov/waterdata/nhdplus-national-data}
}
\usage{
fac_sample
}
\description{
Sample Flow Accumulation
}
\keyword{datasets}
