% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplusTools.R
\docType{data}
\name{fdr_sample}
\alias{fdr_sample}
\title{Sample Flow Direction Raster}
\format{A raster object}
\source{
\url{https://www.epa.gov/waterdata/nhdplus-national-data}
}
\usage{
fdr_sample
}
\description{
Sample Flow Direction
}
\keyword{datasets}
