% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nhd.R
\name{find_huc8}
\alias{find_huc8}
\title{Find WBD HUC 12 unit subsets}
\usage{
find_huc8(AOI = NULL, id = NULL)
}
\arguments{
\item{AOI}{an 'area of interest' object (sf POINT or POLYGON)}

\item{id}{Search for WBD HUC08 features by ID}
}
\value{
simple feature polygon object
}
\description{
Subsets the WBD level 08 features by location (POINT), area (POLYGON), or set of IDs.
}
