% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nhd.R
\name{find_nhd}
\alias{find_nhd}
\title{Find National Hydrography Subsets (Multirealization)}
\usage{
find_nhd(
  AOI = NULL,
  comid = NULL,
  nwis = NULL,
  realization = "flowline",
  streamorder = NULL
)
}
\arguments{
\item{AOI}{an 'area of interest' object (sf POINT or POLYGON)}

\item{comid}{Search for NHD features by COMID}

\item{nwis}{Search for NHD features by colocated NWIS identifiers}

\item{realization}{what realization to return. Default is flowline, options include: outlet, flowline, catchment, and all}

\item{streamorder}{filter returned NHD by streamorder. Returns are for given value and higher. Only usable with AOI, not with comid or nwis inputs}
}
\value{
a single, or list, or simple feature objects
}
\description{
Subsets NHD features by location (POINT), area (POLYGON), or set of COMIDs. Mutli realizations are supported allowing you to query for flowlines, catchments, or outlets.
}
\examples{
 point <- sf::st_sfc(sf::st_point(c(-119.845, 34.4146)), crs = 4326)
 find_nhd(point)
 find_nhd(point, realization = "catchment")
 find_nhd(point, realization = "all")
 find_nhd(comid = 101)
 find_nhd(nwis  = c(11120000, 11120500))
 area <- sf::st_as_sfc(sf::st_bbox(
   c(xmin = -119.8851, xmax =-119.8361, ymax = 34.42439, ymin = 34.40473), crs = 4326))
 find_nhd(area)
 find_nhd(area, realization = "flowline", streamorder = 3)
}
