% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nhd.R
\name{find_outlets}
\alias{find_outlets}
\title{Identify NHD outlets from flowlines}
\usage{
find_outlets(flowlines)
}
\arguments{
\item{flowlines}{a sf LINESTRING or MULTILINESTRING object}
}
\value{
the input flowlines object with geometry changed to POINT objects representing the oulets
}
\description{
Convert flowline NHD realization to outlet POINT objects.
}
