% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nhd.R
\name{find_waterbodies}
\alias{find_waterbodies}
\title{Find WBD HUC08 unit Subsets}
\usage{
find_waterbodies(AOI = NULL, id = NULL)
}
\arguments{
\item{AOI}{an 'area of interest' object (sf POINT or POLYGON)}

\item{id}{Search for WBD HUC08 features by ID}
}
\value{
simple feature polygon object
}
\description{
Subsets NHD waterbody features by location (POINT), area (POLYGON), or set of IDs.
}
