% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNLDI.R
\name{findNLDI}
\alias{findNLDI}
\title{Find Linked Hydro Data via the NLDI}
\usage{
findNLDI(comid = NULL, nwis = NULL, huc12 = NULL, npdes_rad = NULL,
  wqp = NULL, nav = NULL, find = NULL, distance_km = NULL)
}
\arguments{
\item{comid}{an NHD COMID}

\item{nwis}{a USGS NWIS station ID}

\item{huc12}{a HUC12 ID}

\item{npdes_rad}{Facilities that Discharge to Water NHDPlus Indexed Dataset (ID)}

\item{wqp}{a water quality portal location ID}

\item{nav}{where to navigate from the starting point}

\item{find}{what resources to find along the navigation path ("comid", "huc12pp", "npdes_rad", "nwis", "wqp", "basin")}

\item{distance_km}{how far to look along the navigation path (in kilometers)}
}
\value{
a list of sf features
}
\description{
This function provides a formal query to the Network Linked Data Index (NLDI).
There are three tasks that can be accomplished (1) feature discovery (2) NHD Network Tracing (3) Network entity discovery.
One parameter ('comid', 'nwis', 'np_des', or 'wqp') must be supplied to indicate where to start on NHD
The 'nav' parameter describes what direction the network should be traced from the starting point (UT = Upper Tributary, UM = Upper Mainsteam, DD = Down Stream Diversion, DM = Down Stream Mainstem)
Finally, the 'find' parameter, describes what features should be retunred. The 'km' parameter can limit the distance the network is traced.
}
\examples{
\donttest{
# Starting at NWIS 05428500, navigate along the upper mainsteam, and upper tributary
#to find the basin and NWIS sites:

 nldi = findNLDI(nwis = "05428500", nav = c("UT", "UM", "DM"), find = c("nwis", "basin"))

 plot(nldi$basin$geometry, lwd = 2, col = NA, border = 'black')
 plot(nldi$UT$geometry, col = 'blue', add = T)
 plot(nldi$DM$geometry, col = 'purple', add = T)
 plot(nldi$UM$geometry, col = 'red', lwd = 3, add = T)
 plot(nldi$nwissite$geometry, pch = 16, col = 'darkgreen', add = T)
}
}
