% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplusTools.R
\docType{data}
\name{flowline_sample}
\alias{flowline_sample}
\title{Sample Flowlines Walker River}
\format{An sf data.frame}
\source{
\url{https://www.epa.gov/waterdata/nhdplus-national-data}
}
\usage{
flowline_sample
}
\description{
Sample Flowlines
}
\keyword{datasets}
