% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_datasets.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Return RPU or VPU boundaries}
\usage{
get_boundaries(type = "vpu")
}
\arguments{
\item{type}{character. Either "RPU" or "VPU"}
}
\value{
An object of class \code{"sf"}
}
\description{
Return RPU or VPU boundaries
}
