% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oaproc.R
\name{get_elev_along_path}
\alias{get_elev_along_path}
\title{Get Elevation Along Path (experimental)}
\usage{
get_elev_along_path(points, num_pts, res = 1, status = TRUE)
}
\arguments{
\item{points}{sf data.frame containing a point column.}

\item{num_pts}{numeric number of points to retrieve along the cross section.}

\item{res}{integer resolution of 3D Elevation Program data to request.
Must be on of: 1, 3, 5, 10, 30, 60.}

\item{status}{logical}
}
\value{
sf data.frame containing points retrieved. Names include
"id", "distance_m", "elevation_m", "spatial_ref", "geometry",
and ".group". .group tracks which input point each set of output
points belongs to.
}
\description{
Uses a cross section retrieval web services to retrieve elevation
 along a path.
}
