% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hydro.R
\name{get_gagesII}
\alias{get_gagesII}
\title{Find GAGESII Features}
\usage{
get_gagesII(AOI = NULL, id = NULL)
}
\arguments{
\item{AOI}{sf (MULTI)POINT or (MULTI)POLYGON. An 'area of interest' can
be given as either a location (sf POINT) or area (sf POLYGON)
in any coordinate reference system.}

\item{id}{Search for gagesII locations by NWIS siteID}
}
\value{
simple feature polygon object
}
\description{
Subsets the gagesII dataset by location (POINT),
area (POLYGON), or set of IDs.
}
