% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhdplushr.R
\name{get_hr_data}
\alias{get_hr_data}
\title{Get NHDPlus HiRes Data}
\usage{
get_hr_data(gdb, layer = NULL, min_size_sqkm = NULL, simp = NULL,
  proj = NULL, keep_cols = NULL, rename = TRUE)
}
\arguments{
\item{gdb}{character path to geodatabase to get data from.}

\item{layer}{character layer name from geodatabase found with \link{sf::st_layers}}

\item{min_size_sqkm}{numeric minimum basin size to be included in the output}

\item{simp}{numeric simplification tolerance in units of projection}

\item{proj}{a projection specification compatible with \link{sf::st_crs}}

\item{keep_cols}{character vector of column names to keep in the output. If NULL,
all will be kept.}
}
\description{
Get NHDPlus HiRes Data
}
