% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hydro.R
\name{get_huc8}
\alias{get_huc8}
\title{Find WBD HUC 08 unit subsets}
\usage{
get_huc8(AOI = NULL, id = NULL)
}
\arguments{
\item{id}{WBD HUC08 ID(s)}
}
\value{
simple feature polygon object
}
\description{
Subsets the WBD level 08 features by location (POINT),
area (POLYGON), or set of IDs.
}
