% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_nhdplus.R
\name{get_hydro_location}
\alias{get_hydro_location}
\title{Get hydro location}
\usage{
get_hydro_location(indexes, flowpath)
}
\arguments{
\item{indexes}{data.frame as output from \link{get_flowline_index}.}

\item{flowpath}{data.frame with two three columns, COMID, FromMeas, and ToMeas.
The first should join to the COMID field of the indexes and the second
should be linear geometry.}
}
\description{
given a flowline index, returns the hydrologic location (point)
along the specific linear element referenced by the index.
}
\examples{
source(system.file("extdata", "sample_flines.R", package = "nhdplusTools"))

indexes <- get_flowline_index(sample_flines,
                   sf::st_sfc(sf::st_sfc(list(sf::st_point(c(-76.86934, 39.49328)),
                                              sf::st_point(c(-76.91711, 39.40884)),
                                              sf::st_point(c(-76.88081, 39.36354))),
                              crs = 4326)))

get_hydro_location(indexes, sample_flines)

}
