% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked_points.R
\name{get_linked_points}
\alias{get_linked_points}
\title{Get linked points}
\usage{
get_linked_points(hu_lp, net, wbd, exclude, cores = NA)
}
\arguments{
\item{hu_lp}{data.frame as returned by `match_levelpaths`}

\item{net}{sf data.frame ...}

\item{wbd}{sf data.frame ...}

\item{exclude}{character vector of HUs to exclude from consideration}

\item{cores}{numeric (optional) number of cores to use in paralel evaluation

This function is preliminary and subject to revision. It has been tested thoroughly
but complete methods description have not yet been published.}
}
\description{
Given matched levelpaths, finds outlets of catchments along level paths.
}
\examples{
library(dplyr)
library(sf)

source(system.file("extdata/new_hope_data.R", package = "nhdplusTools"))
suppressWarnings(net_prep <- prepare_nhdplus(new_hope_flowline, 20, 0, 10) \%>\%
                   left_join(select(new_hope_flowline, COMID, DnLevelPat, AreaSqKM), by = "COMID") \%>\%
                   st_sf() \%>\%
                   group_by(LevelPathI) \%>\%
                   arrange(Hydroseq) \%>\%
                   mutate(DnLevelPat = DnLevelPat[1]) \%>\%
                   ungroup())

net_prep["denTotalAreaSqKM"] <-
  calculate_total_drainage_area(select(st_set_geometry(net_prep, NULL),
                                       ID = COMID, toID = toCOMID,
                                       area = AreaSqKM))

wbd <- select(new_hope_wbd, HUC12 = HUC_12, TOHUC = HU_12_DS) \%>\%
  st_transform(st_crs(net_prep))

net_prep <- st_join(net_prep, wbd) \%>\%
  st_set_geometry(NULL)

outlet_comid <- 8897784

hu_lp <- match_levelpaths(net_prep, outlet_comid)

linked_points <- get_linked_points(hu_lp, new_hope_flowline, wbd, exclude = c())

lps <- filter(new_hope_flowline, LevelPathI \%in\% linked_points$LevelPathI)

plot(st_geometry(new_hope_flowline), lwd = 0.4, col = "blue")
plot(st_geometry(lps), col = "blue", add = TRUE)
plot(linked_points$geometry, col = "red", pch = 19, add = TRUE)
plot(st_transform(new_hope_wbd$geom,
                  st_crs(linked_points)), lwd = 1.5, add = TRUE)
}
