% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nldi.R
\name{get_nldi_characteristics}
\alias{get_nldi_characteristics}
\title{(DEPRECATED) Get Catchment Characteristics}
\usage{
get_nldi_characteristics(nldi_feature, type = "local")
}
\arguments{
\item{nldi_feature}{list with names `featureSource` and `featureID` where
`featureSource` is derived from the "source" column of  the response of
\link[dataRetrieval]{get_nldi_sources} and the `featureID` is a known identifier
from the specified `featureSource`.}

\item{type}{character "all", "local", "total", or "divergence_routed".}
}
\value{
data.frame containing requested characteristics
}
\description{
This functionality id deprecated and will be removed in the near future.

Please use \link{get_catchment_characteristics} instead.
}
