% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hydro.R
\name{get_nwis}
\alias{get_nwis}
\title{Discover USGS NWIS Stream Gages}
\usage{
get_nwis(AOI = NULL)
}
\arguments{
\item{AOI}{sf (MULTI)POINT or (MULTI)POLYGON. An 'area of interest' can
be given as either a location (sf POINT) or area (sf POLYGON)
in any coordinate reference system.}
}
\value{
sf object
}
\description{
Returns a POINT feature class of active, stream network,
NWIS gages for an Area of Interest. If a POINT is given as the AOI,
a 20km2 bounding box is searched and the nearest feature (euclidian)
is returned. Data is accessed through the NWIS web portal.
}
