% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_lengths.R
\name{get_path_lengths}
\alias{get_path_lengths}
\title{Get Path Lengths (DEPRECATED)}
\usage{
get_path_lengths(outlets, network, cores = 1, status = FALSE)
}
\arguments{
\item{outlets}{vector of IDs from data.frame}

\item{network}{data.frame with ID, toID, and lengthkm attributes.}

\item{cores}{integer number of cores to use for parallel computation.}

\item{status}{logical print status and progress bars?}
}
\value{
data.frame containing the distance between pairs of network outlets.
For a network with one terminal outlet, the data.frame will have
`nrow(network)^2` rows.
}
\description{
Given a network and set of IDs, finds path lengths between all
identified flowpath outlets. This algorithm finds distance between outlets
regardless of flow direction.
}
