% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_lengths.R
\name{get_path_members}
\alias{get_path_members}
\title{Get Path Members (DEPRECATED)}
\usage{
get_path_members(outlets, network, cores = 1, status = FALSE)
}
\arguments{
\item{outlets}{vector of IDs from data.frame}

\item{network}{data.frame with ID, toID, and lengthkm attributes.}

\item{cores}{integer number of cores to use for parallel computation.}

\item{status}{logical print status and progress bars?}
}
\value{
list of lists containing flowpath identifiers along path that connect
outlets.
}
\description{
Given a network and set of IDs, finds paths between all
identified flowpath outlets. This algorithm finds members between outlets
regardless of flow direction.
}
