% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_pathlength}
\alias{get_pathlength}
\title{Get Path Length}
\usage{
get_pathlength(x)
}
\arguments{
\item{x}{data.frame with ID, toID, length columns.}
}
\value{
data.frame containing pathlength for each ID
}
\description{
Generates the main path length to a basin's terminal path.
}
