% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{get_pfaf}
\alias{get_pfaf}
\title{Get Pfafstetter Codes (DEPRECATED)}
\usage{
get_pfaf(x, max_level = 2, status = FALSE)
}
\arguments{
\item{x}{sf data.frame with ID, toID, totda, outletID, topo_sort,
and levelpath attributes.}

\item{max_level}{integer number of pfaf levels to attempt to calculate.
If the network doesn't have resolution to support the desired level,
unexpected behavior may occur.}

\item{status}{boolean print status or not}
}
\value{
data.frame with ID and pfaf columns.
}
\description{
Determines Pfafstetter codes for a dendritic network with
total drainage area, levelpath, and topo_sort attributes.
}
