% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_watershed_delineation_nhdplushr.R
\name{get_split_catchments}
\alias{get_split_catchments}
\title{get_split_catchments}
\usage{
get_split_catchments(points)
}
\arguments{
\item{points}{an sf object of type POINT or MULTIPOINT containing the the locations
of desired catchment outlets}
}
\value{
an sfc object of catchment polygons for each point
}
\description{
Get catchments using USGS webservice get_split_catchment
NOTE: Sometimes USGS webservice fails, and it uses nhdplusv2 (not high res) so use if this is not
recommended.
}
