% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_terminal}
\alias{get_terminal}
\title{Get Terminal ID (DEPRECATED)}
\usage{
get_terminal(x, outlets)
}
\arguments{
\item{x}{two column data.frame with IDs and toIDs. Names are ignored.}

\item{outlets}{IDs of outlet flowlines}
}
\value{
data.frame containing the terminal ID for each outlet
}
\description{
Get the ID of the basin outlet for each flowline.
This function has been deprecated in favor of get_sorted.
}
