% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_watershed_delineation_nhdplushr.R
\name{get_upstream_catchment_from_nhdplushr}
\alias{get_upstream_catchment_from_nhdplushr}
\title{get_upstream_catchment_from_nhdplushr}
\usage{
get_upstream_catchment_from_nhdplushr(points, flowlines, catchments, crs = crs)
}
\arguments{
\item{points}{an sf object of type POINT or MULTIPOINT containing the the locations
of desired catchment outlets}

\item{flowlines}{an sf object containing nhdplushr flowlines}

\item{catchments}{an sf object conatining catchment geometries for each nhdplushr flowline}

\item{crs}{the coordinate reference system for geoprocessing. Defaults to epsg:4326}
}
\value{
an sfc object of catchment geometries for each point
}
\description{
Use nhdplus catchments for each flowline and union them together - this might
take a while but will likely result in better catchments than from the USGS
webservice in get_split_catchments.
NOTE: TO use this we need to also to set layers = c("NHDFlowline", "NHDPlusCatchment")
in get_nhdplushr in the main function
}
