% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_watershed_delineation_nhdplushr.R
\name{get_upstream_flowlines}
\alias{get_upstream_flowlines}
\title{get_upstream_flowlines}
\usage{
get_upstream_flowlines(points, flowlines)
}
\arguments{
\item{points}{an sf object of type POINT or MULTIPOINT}

\item{flowlines}{an sf object containing nhdplushr flowlines}
}
\value{
an sf object with a list column containing all the upstream COMIDs/hydroseq
}
\description{
Function to get nearest from points COMID/hydroseq and all upstream
}
