% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{get_UT}
\alias{get_UT}
\title{Navigate Upstream with Tributaries}
\usage{
get_UT(network, comid, distance = NULL)
}
\arguments{
\item{network}{data.frame NHDPlus flowlines including at a minimum:
COMID, Pathlength, LENGTHKM, and Hydroseq.}

\item{comid}{integer Identifier to start navigating from.}

\item{distance}{numeric distance in km to limit how many COMIDs are
returned. The COMID that exceeds the distance specified is returned.}
}
\value{
integer vector of all COMIDs upstream with tributaries of the
starting COMID.
}
\description{
Traverse NHDPlus network upstream with tributaries
}
