% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaa.R
\name{get_vaa_path}
\alias{get_vaa_path}
\title{File path to value added attribute (vaa) Cache}
\usage{
get_vaa_path()
}
\value{
file.path character
}
\description{
nhdplusTools will download and cache an `fst` file with
NHDPlusV2 attribute data sans geometry. This function returns the
file path to the cached file. Will use the user cache dir indicated
by \link[rappdirs]{user_cache_dir}.
}
\examples{
get_vaa_path()
}
