% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hydro.R
\name{get_waterbodies}
\alias{get_waterbodies}
\title{Find NHD Water Bodies}
\usage{
get_waterbodies(AOI = NULL, id = NULL)
}
\arguments{
\item{id}{Search for NHD Water Bodies by COMID}
}
\value{
simple feature polygon object
}
\description{
Subsets NHD waterbody features by location (POINT),
area (POLYGON), or set of IDs.
}
