% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_nhdplus.R
\name{get_waterbody_index}
\alias{get_waterbody_index}
\title{Get Waterbody Index}
\usage{
get_waterbody_index(waterbodies, points, flines = NULL, search_radius = NULL)
}
\arguments{
\item{waterbodies}{sf data.frame of type POLYGON or MULTIPOLYGON including
COMID attributes.}

\item{points}{sfc of type POINT}

\item{flines}{sf data.frame of type LINESTRING or MULTILINESTRING including
COMID, WBAREACOMI, and Hydroseq attributes}

\item{search_radius}{units class with a numeric value indicating how far to
search for a waterbody boundary in units of provided projection. Set units with
\link[units]{set_units}.}
}
\value{
data.frame with two columns, COMID, in_wb_COMID, near_wb_COMID,
near_wb_dist, and outlet_fline_COMID. Distance is in units of provided projection.
}
\description{
given an sf point geometry column, return waterbody id, and
COMID of dominant artificial path
}
