% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_hydrogeofetch.R
\name{hydrogeofetch_cache_settings}
\alias{hydrogeofetch_cache_settings}
\title{hydrogeofetch cache settings}
\usage{
hydrogeofetch_cache_settings(mode = NULL, timeout = NULL)
}
\arguments{
\item{mode}{character 'memory' or 'filesystem'}

\item{timeout}{numeric number of seconds until caches invalidate}
}
\value{
list containing settings at time of calling. If inputs are
NULL, current settings. If settings are altered, previous setting values.
}
\description{
Provides an interface to adjust hydrogeofetch `memoise` cache.

Mode and timeout can also be set using environment variables.
`HYDROGEOFETCH_MEMOISE_CACHE` and `HYDROGEOFETCH_MEMOISE_TIMEOUT` are
used unless overridden with this function. The old `NHDPLUSTOOLS_*`
names are still recognized as a fallback.
}
