% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_hydrogeofetch.R
\name{hydrogeofetch_data_dir}
\alias{hydrogeofetch_data_dir}
\title{Get or set hydrogeofetch data directory}
\usage{
hydrogeofetch_data_dir(dir = NULL)
}
\arguments{
\item{dir}{path of desired data directory}
}
\value{
character path of data directory (silent when setting)
}
\description{
if left unset, will return the user data dir
as returned by `tools::R_user_dir` for this package.
}
\examples{
hydrogeofetch_data_dir()

hydrogeofetch_data_dir("demo")

hydrogeofetch_data_dir(tools::R_user_dir("hydrogeofetch"))

}
