% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_levelpaths.R
\name{match_levelpaths}
\alias{match_levelpaths}
\title{Match Level Paths}
\usage{
match_levelpaths(fline_hu, start_comid, add_checks = FALSE)
}
\arguments{
\item{fline_hu}{sf data.frame flowlines intersected with hydrologic units containing
COMID, Hydroseq, LevelPathI, DnLevelPat, denTotalAreaSqKM, HUC12, TOHUC attributes.}

\item{start_comid}{integer COMID to start search from.}

\item{add_checks}{boolean if TRUE, checks for toHUC errors are added.}
}
\description{
Attempts to match dendritic hydrologic unit networks to level paths.
}
\details{
Match level paths compares the set of hydrologic units found through
spatial intersection to the path identified by navigating the TOHUC codes from
headwater to outlet and attempts of reconcile anomolies found.

This function is preliminary and subject to revision. It has been tested thoroughly
but complete methods description have not yet been published.
}
\examples{
net_prep <- readRDS(system.file("extdata/match_levelpaths_2279159.rds", package = "nhdplusTools"))
match_levelpaths(net_prep, 2279159, add_checks = TRUE)

}
