% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_nhdplusTools.R
\name{nhdplus_path}
\alias{nhdplus_path}
\title{NHDPlus Data Path}
\usage{
nhdplus_path(path = NULL, warn = FALSE)
}
\arguments{
\item{path}{character path ending in .gdb or .gpkg}

\item{warn}{boolean controls whether warning an status messages are printed}
}
\value{
0 (invisibly) if set successfully, character path if no input.
}
\description{
Allows specification of a custom path to a source dataset.
Typically this will be the national seamless dataset in
geodatabase or geopackage format.
}
\examples{
nhdplus_path("/data/NHDPlusV21_National_Seamless.gdb")

nhdplus_path("/data/NHDPlusV21_National_Seamless.gdb", warn=FALSE)

nhdplus_path()

}
