% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplus_refactor.R
\name{nhdplus_refactor}
\alias{nhdplus_refactor}
\title{nhdplus refactor}
\usage{
nhdplus_refactor(nhdplus_flines, split_flines_meters, split_flines_cores,
  collapse_flines_meters, collapse_flines_main_meters, out_collapsed,
  out_reconciled, three_pass = FALSE)
}
\arguments{
\item{nhdplus_flines}{data.frame nhdplus flowloines}

\item{split_flines_meters}{numeric split threshold in meters}

\item{split_flines_cores}{numeric number of cores to use for split function}

\item{collapse_flines_meters}{numeric collapse threshold in meters}

\item{collapse_flines_main_meters}{numeric collapse threshold for mainstems in meters}

\item{out_collapsed}{character geopackage path to write collapsed to (one layer per gpkg)}

\item{out_reconciled}{character geopackage path to write reconciled to (one layer per gpkg)}

\item{three_pass}{boolean perform three collapse passes or just one}
}
\description{
refactors nhdplud flowlines
}
