% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplus_refactor_2.R
\name{nhdplus_refactor_2}
\alias{nhdplus_refactor_2}
\title{nhdplus refactor 2}
\usage{
nhdplus_refactor_2(nhdplus_flines, split_flines_meters, split_flines_cores,
  collapse_flines_meters, collapse_flines_mainstem_meters, out_collapsed,
  out_reconciled)
}
\description{
refactors nhdplud flowlines
}
