% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_nhdplusTools.R
\name{nhdplusTools_data_dir}
\alias{nhdplusTools_data_dir}
\title{Get or set nhdplusTools data directory}
\usage{
nhdplusTools_data_dir(dir = NULL)
}
\arguments{
\item{dir}{path of desired data directory}
}
\value{
character path of data directory (silent when setting)
}
\description{
if left unset, will return the user data dir
as returned by `tools::R_user_dir` for this package.
}
\examples{
nhdplusTools_data_dir()

nhdplusTools_data_dir("demo")

nhdplusTools_data_dir(tools::R_user_dir("nhdplusTools"))

}
