% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nhdplus.R
\name{plot_nhdplus}
\alias{plot_nhdplus}
\title{Plot NHDPlus}
\usage{
plot_nhdplus(
  outlets = NULL,
  bbox = NULL,
  streamorder = NULL,
  nhdplus_data = NULL,
  gpkg = NULL,
  plot_config = NULL,
  basemap = "Esri.NatGeoWorldMap",
  zoom = NULL,
  add = FALSE,
  actually_plot = TRUE,
  overwrite = TRUE,
  flowline_only = NULL,
  cache_data = NULL
)
}
\arguments{
\item{outlets}{list of nldi outlets. Other inputs are coerced into nldi outlets, see details.}

\item{bbox}{object of class bbox with a defined crs. See examples.}

\item{streamorder}{integer only streams of order greater than or equal will be returned}

\item{nhdplus_data}{geopackage containing source nhdplus data (omit to download)}

\item{gpkg}{path and file with .gpkg ending. If omitted, no file is written.}

\item{plot_config}{list containing plot configuration, see details.}

\item{basemap}{character indicating which basemap type to use. Chose from:
\link[maptiles]{get_tiles}.}

\item{zoom}{integer passed on to \link[maptiles]{get_tiles}. This value will
override the default set by the package.}

\item{add}{boolean should this plot be added to an already built map.}

\item{actually_plot}{boolean actually draw the plot? Use to get data subset only.}

\item{overwrite}{passed on the \link{subset_nhdplus}.}

\item{flowline_only}{boolean only subset and plot flowlines only, default=FALSE}

\item{cache_data}{character path to rds file where all plot data can be cached.
If file doesn't exist, it will be created. If set to FALSE, all caching will
be turned off -- this includes basemap tiles.}
}
\value{
data.frame plot data is returned invisibly in NAD83 Lat/Lon.
}
\description{
Given a list of outlets, get their basin boundaries and network and return a plot in
EPSG:3857 Web Mercator Projection.
}
\details{
plot_nhdplus supports several input specifications. An unexported function "as_outlet"
is used to convert the outlet formats as described below.
\enumerate{
  \item if outlets is omitted, the bbox input is required and all nhdplus data
  in the bounding box is plotted.
  \item If outlets is a list of integers, it is assumed to be NHDPlus IDs (comids)
  and all upstream tributaries are plotted.
  \item if outlets is an integer vector, it is assumed to be all NHDPlus IDs (comids)
  that should be plotted. Allows custom filtering.
  \item If outlets is a character vector, it is assumed to be NWIS site ids.
  \item if outlets is a list containing only characters, it is assumed to be a list
  of nldi features and all upstream tributaries are plotted.
  \item if outlets is a data.frame with point geometry, a point in polygon match
  is performed and upstream with tributaries from the identified catchments is plotted.
}

The \code{plot_config} parameter is a list with names "basin", "flowline", "outlets",
 "network_wtbd", and "off_network_wtbd".
The following shows the defaults that can be altered.
\enumerate{
  \item basin \preformatted{
    list(lwd = 1, col = NA, border = "black")}
  \item flowline \preformatted{
    list(lwd = 1, col = "blue")}
  \item outlets \preformatted{
   list(default = list(col = "black", border = NA, pch = 19, cex = 1),
        nwissite = list(col = "grey40", border = NA, pch = 17, cex = 1),
        huc12pp = list(col = "white", border = "black", pch = 22, cex = 1),
        wqp = list(col = "red", border = NA, pch = 20, cex = 1))}
  \item network_wtbd  \code{list(lwd = 1, col = "lightblue", border = "black")}
  \item off_network_wtbd  \code{list(lwd = 1, col = "darkblue", border = "black")}

}

If adding additional layers to the plot, data must be projected to EPSG:3857 with
`sf::st_transform(x, 3857)` prior to adding to the plot.
}
