% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile_catchments}
\alias{reconcile_catchments}
\title{Reconcile Catchments}
\usage{
reconcile_catchments(catchment, fline_ref, fline_rec, fdr, fac, para = 2)
}
\arguments{
\item{catchment}{sf data.frame NHDPlus Catchment or CatchmentSP for included COMIDs}

\item{fline_ref}{sf data.frame flowlines as returned by
\code{\link{refactor_nhdplus}} and \code{\link{reconcile_collapsed_flowlines}}}

\item{fline_rec}{sf data.frame flowlines as returned by
\code{\link{reconcile_collapsed_flowlines}} and
\code{\link{reconcile_collapsed_flowlines}}}

\item{fdr}{raster D8 flow direction}

\item{fac}{raster flow accumulation}

\item{para}{integer numer of cores to use for parallel execution}
}
\value{
Catchments that have been split and collapsed according to input flowlines
}
\description{
Reconciles catchments according to the output of
\code{\link{reconcile_collapsed_flowlines}} and \code{\link{refactor_nhdplus}}
}
\details{
Note that all inputs must be passed in the same projection.
}
\seealso{
The \code{\link{refactor_nhdplus}} function implements a complete
workflow using `reconcile_collapsed_flowlines()` and can be used in prep
for this function.
}
