% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile_collapsed_flowlines.R
\name{reconcile_collapsed_flowlines}
\alias{reconcile_collapsed_flowlines}
\title{Reconcile collapsed flowlines}
\usage{
reconcile_collapsed_flowlines(flines, geom = NULL, id = "COMID")
}
\value{
reconciled flines
}
\description{
Reconciles output of collapse_flowlines giving a unique ID to each new unit and providing a mapping to NHDPlus COMIDs.
}
