% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile_downstream}
\alias{reconcile_downstream}
\title{Reconcile removed flowlines in the downstream direction}
\usage{
reconcile_downstream(flines, remove_fun, remove_problem_headwaters = FALSE)
}
\arguments{
\item{flines}{data.frame with COMID, toCOMID, LENGTHKM, and TotDASqKM columns}

\item{remove_fun}{a function that will return boolean indicating which flines should be removed}

\item{remove_problem_headwaters}{whether to mark headwaters upstream of confluences as unable to remove}
}
\value{
flines data.frame with short flowlines merged downstream
}
\description{
... TBD ...
}
