% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile_removed_flowlines}
\alias{reconcile_removed_flowlines}
\title{Reconcile removed flowlines}
\usage{
reconcile_removed_flowlines(flines, reroute_set, removed, original_fline_atts,
  normalize_mainstems = FALSE)
}
\arguments{
\item{flines}{data.frame ...}

\item{reroute_set}{boolean ...}

\item{removed}{data.frame ...}

\item{original_fline_atts}{data.frame ...}

\item{normalize_mainstems}{boolean ...}
}
\value{
reconciled flines
}
\description{
Reconciles removed/missing flowlines from the network.
}
