% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplus_refactor.R
\name{refactor_nhdplus}
\alias{refactor_nhdplus}
\title{nhdplus refactor}
\usage{
refactor_nhdplus(nhdplus_flines, split_flines_meters, split_flines_cores,
  collapse_flines_meters, collapse_flines_main_meters, out_collapsed,
  out_reconciled, three_pass = FALSE, purge_non_dendritic = TRUE,
  warn = TRUE)
}
\arguments{
\item{nhdplus_flines}{data.frame nhdplus flowloines}

\item{split_flines_meters}{numeric split threshold in meters}

\item{split_flines_cores}{numeric number of cores to use for split function}

\item{collapse_flines_meters}{numeric collapse threshold in meters}

\item{collapse_flines_main_meters}{numeric collapse threshold for mainstems in meters}

\item{out_collapsed}{character geopackage path to write collapsed to (one layer per gpkg)}

\item{out_reconciled}{character geopackage path to write reconciled to (one layer per gpkg)}

\item{three_pass}{boolean perform three collapse passes or just one}

\item{purge_non_dendritic}{boolean passed on to prepare_nhdplus}

\item{warn}{boolean controls whether warning an status messages are printed}
}
\description{
refactors nhdplus flowlines
}
\examples{
nhdplus_flowlines <- sf::st_zm(sample_flines)
refactor_nhdplus(nhdplus_flines = nhdplus_flowlines,
                 split_flines_meters = 2000,
                 split_flines_cores = 3,
                 collapse_flines_meters = 500,
                 collapse_flines_main_meters = 500,
                 out_collapsed = "temp.gpkg",
                 out_reconciled = "temp_rec.gpkg",
                 three_pass = TRUE,
                 purge_non_dendritic = FALSE,
                 warn = FALSE)
unlink("temp.gpkg")
unlink("temp_rec.gpkg")

}
