% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplusTools.R
\name{rename_geometry}
\alias{rename_geometry}
\title{rename_geometry}
\usage{
rename_geometry(g, name)
}
\arguments{
\item{g}{sf data.table}

\item{name}{character name to be used for geometry}
}
\description{
correctly renames the geometry column
of a sf object.
}
