% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_datasets.R
\docType{data}
\name{rpu}
\alias{rpu}
\title{RPU Boundaries
Raster Processing Unit boundaries}
\format{
An object of class \code{"sf"}
}
\usage{
rpu
}
\description{
RPU Boundaries
Raster Processing Unit boundaries
}
\keyword{data}
