% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_flines}
\alias{sample_flines}
\title{Sample Flowlines from the Petapsco River.}
\format{sf data.frame}
\source{
\url{https://www.epa.gov/waterdata/nhdplus-national-data}
}
\usage{
sample_flines
}
\description{
A sample set of flowlines.
}
\keyword{datasets}
