% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_catchment.R
\name{split_catchment}
\alias{split_catchment}
\title{Split Catchments}
\usage{
split_catchment(catchment, fline, fdr, fac)
}
\arguments{
\item{catchment}{sf data.frame with one catchment}

\item{fline}{sf data.frame with two or more flowline segments in
upstream downstream order.}

\item{fdr}{raster a flow direction raster that fully covers the catchment}

\item{fac}{raster a flow accumulation raster that fuller covers the catchment}
}
\value{
Split catchments as an sfc geometry.
}
\description{
A catchment splitting algorithm that works with a D8
flow direction grid and the output of nhdplus_refactor. See Vignette
for examples.
}
