% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_flowlines.R
\name{split_flowlines}
\alias{split_flowlines}
\title{split flowlines}
\usage{
split_flowlines(flines, max_length)
}
\value{
all the flowlines with long ones replaced with split ones.
}
\description{
Splits flowlines longer than a given threshold into the minimum number of pieces to all be under the given threshold.
}
