% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_lines.R
\name{split_lines_2}
\alias{split_lines_2}
\title{split lines 2}
\usage{
split_lines_2(input_lines, max_length, id = "ID")
}
\arguments{
\item{input_lines}{data.frame of class sf with LINESTRING sfc column.}

\item{max_length}{maximum segment length to return}

\item{id}{name of ID column in data.frame}
}
\value{
only the split lines.
}
\description{
Splits lines longer than a given threshold into the
minimum number of pieces to all be under the given threshold.
Does not use lwgeom and may not work in all cases.
}
