% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{stage_national_data}
\alias{stage_national_data}
\title{Stage NHDPlus National Data}
\usage{
stage_national_data(include = c("attribute", "flowline", "catchment"),
  output_path = NULL, nhdplus_data_path = NULL)
}
\arguments{
\item{include}{character vector containing one or more of:
"attributes", "flowline", "catchment".}

\item{output_path}{character path to save the output to defaults
to the directory of the nhdplus_data_path.}

\item{nhdplus_data_path}{character path to the .gpkg or .gdb
containing the national seamless dataset. Not required if
\code{\link{nhdplus_path}} has been set.}
}
\value{
list containing paths to the .rds files.
}
\description{
Breaks down the national geo database into a collection
of quick to access R binary files.
}
\details{
"attributes" will save `NHDFlowline_Network` attributes
as a seperate data.frame without the geometry. The others will save
the `NHDFlowline_Network` and `CatchmentSP` as sf data.frames with
superfluous Z information dropped.

The returned list of paths is also added to the nhdplusTools_env
as "national_data".
}
\examples{
\dontrun{
stage_national_data()

stage_national_data(include = c("attributes", "flowlines","catchments"))
}

}
