% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_nhdplus.R
\name{subset_nhdplus}
\alias{subset_nhdplus}
\title{Subset NHDPlus}
\usage{
subset_nhdplus(
  comids = NULL,
  output_file = NULL,
  nhdplus_data = NULL,
  bbox = NULL,
  simplified = TRUE,
  overwrite = FALSE,
  return_data = TRUE,
  status = TRUE,
  flowline_only = NULL,
  streamorder = NULL,
  out_prj = 4269
)
}
\arguments{
\item{comids}{integer vector of COMIDs to include.}

\item{output_file}{character path to save the output to defaults
to the directory of the nhdplus_data.}

\item{nhdplus_data}{character path to the .gpkg or .gdb containing
the national seamless database, a subset of NHDPlusHR,
or "download" to use a web service to download NHDPlusV2.1 data.
Not required if \code{\link{nhdplus_path}} has been set or the default
has been adopted. See details for more.}

\item{bbox}{object of class "bbox" as returned by sf::st_bbox in Latitude/Longitude.
If no CRS is present, will be assumed to be in WGS84 Latitude Longitude.}

\item{simplified}{boolean if TRUE (the default) the CatchmentSP layer
will be included. Not relevant to the "download" option or NHDPlusHR data.}

\item{overwrite}{boolean should the output file be overwritten}

\item{return_data}{boolean if FALSE path to output file is returned silently otherwise
data is returned in a list.}

\item{status}{boolean should the function print status messages}

\item{flowline_only}{boolean WARNING: experimental
if TRUE only the flowline network and attributes will be returned}

\item{streamorder}{integer only streams of order greater than or equal will be downloaded.
Not implemented for local data.}

\item{out_prj}{character override the default output CRS of NAD83 lat/lon (EPSG:4269)}
}
\value{
character path to the saved subset geopackage
}
\description{
Saves a subset of the National Seamless database or other
nhdplusTools compatible data based on a specified collection of COMIDs.
This function uses \code{\link{get_nhdplus}} for the "download" data
source but returns data consistent with local data subsets in a subset
file.
}
\details{
This function relies on the National Seamless Geodatabase or Geopackage.
It can be downloaded
\href{https://www.epa.gov/waterdata/nhdplus-national-data}{here.}

The "download" option of this function should be considered preliminary
and subject to revision. It does not include as many layers and may not
be available permanently.
}
