% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaa.R
\name{vaa_path}
\alias{vaa_path}
\title{File path to VAA Cache}
\usage{
vaa_path()
}
\value{
file.path (character)
}
\description{
nhdplusTools will download and cache an fst file with NHDPlusV2 attribute data sans geometry. This function returns the file path to the cached file.
}
