% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_datasets.R
\docType{data}
\name{vpu}
\alias{vpu}
\title{VPU Boundaries
Vector Processing Unit boundaries}
\format{
An object of class \code{"sf"}
}
\usage{
vpu
}
\description{
VPU Boundaries
Vector Processing Unit boundaries
}
\keyword{data}
