% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{walker_catchment}
\alias{walker_catchment}
\title{Sample Catchments from the Walker River.}
\format{sf data.frame}
\source{
\url{https://www.epa.gov/waterdata/nhdplus-national-data}
}
\usage{
walker_catchment
}
\description{
Sample Catchments
}
\keyword{datasets}
