% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{walker_fac}
\alias{walker_fac}
\title{Sample Flow Accumulation Raster}
\format{raster object}
\source{
\url{https://www.epa.gov/waterdata/nhdplus-national-data}
}
\usage{
walker_fac
}
\description{
Sample Flow Accumulation
}
\keyword{datasets}
