% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{walker_fdr}
\alias{walker_fdr}
\title{Sample Flow Direction Raster from the Walker River.}
\format{raster object}
\source{
\url{https://www.epa.gov/waterdata/nhdplus-national-data}
}
\usage{
walker_fdr
}
\description{
Sample Flow Direction
}
\keyword{datasets}
