% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{walker_flowline}
\alias{walker_flowline}
\title{Sample Flowlines Walker River}
\format{sf data.frame}
\source{
\url{https://www.epa.gov/waterdata/nhdplus-national-data}
}
\usage{
walker_flowline
}
\description{
Sample Flowlines
}
\keyword{datasets}
