% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainstems.R
\name{add_mainstems}
\alias{add_mainstems}
\title{Add mainstem identifiers}
\usage{
add_mainstems(x, join_col = NULL, join_col_type = NULL)
}
\arguments{
\item{x}{data.frame or sf containing an identifier column joinable to
NHDPlusV2 or NHDPlusHR.}

\item{join_col}{character name of the identifier column in \code{x}.
Detected automatically from comid, featureid, or nhdplusid (any case)
if not provided.}

\item{join_col_type}{character one of "nhdpv2" or "nhdphr". Detected
automatically from \code{join_col} if not provided.}
}
\value{
\code{x} with mainstem_uri and mainstemid columns added.
}
\description{
Joins geoconnex mainstem identifiers onto a table that
contains NHDPlusV2 (comid/featureid) or NHDPlusHR (nhdplusid)
identifiers using the ref_rivers lookup table. See
\link{download_mainstem_lookup} for source data details.
}
\examples{
\donttest{
add_mainstems(data.frame(comid = c(2804607, 2804621)))
}
}
