% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_catchments.R
\name{aggregate_catchments}
\alias{aggregate_catchments}
\title{Aggregate Catchments}
\usage{
aggregate_catchments(fline_rec, cat_rec, outlets, da_thresh = NA,
  only_larger = FALSE)
}
\arguments{
\item{fline_rec}{sf data.frame Reconciled flowlines as generated by `refactor_nhdplus`}

\item{cat_rec}{sf data.frame Reconciled catchments as genberated by `reconcile_catchments`}

\item{outlets}{data.frame with "ID" and "type" columns. "ID" must be identifiers from
fline and cat data.frames. "type" should be "outlet", or "terminal".
"outlet" will include the specified ID.
"terminal" will be treated as a terminal node with nothing downstream.
This parameter may change.}

\item{da_thresh}{numeric Defaults to NA. A threshold total drainage area in the units of the TotDASqKM
field of the fline_rec data.frame. When automatically adding confluences to make
the network valid, tributary catchments under this threshold will be lumped with
the larger tributaries rather than being added to the set of output catchments.}

\item{only_larger}{boolean Defaults to TRUE. If TRUE when adding confluences to
make the network valid, only tributaries larger than the one with an upstream
outlet will be added. e.g. if a tributary is required in the model this will
add main stems that the tributary contributes to. Note that the NHDPlus treats
divergences as part of the main stem, so the da_thresh may still be needed to
eliminate small tributary catchments introduced by divergences near confluences.}
}
\description{
Aggregates catchments according to a set of outlet catchments.
}
\details{
This function operates on the catchment network as a node-edge graph.
The outlet types are required to ensure that graph searches start from the
appropriate nodes and includes the appropriate catchments. Outlets such as gages
should be treated as "outlet" outlets.
While it may be possible for the algorithm to determine terminal outlets, at this
time, it is required that they be specified explicitely as "terminal" outlet types.

The function checks supplied catchment outlets to make sure they connect to downstream
catchments. Checks verify that the outlet of the levelpath of each supplied outlet is
in the supplied outlet set. If the outlet of a levelpath is not in the supplied set, it
is added along with other catchments that contribute to the same receiving catchment.
These checks ensure that all output catchments have one and only one input and output
nexus and that all catchments are well-connected.
}
\examples{
source(system.file("extdata", "walker_data.R", package = "nhdplusTools"))
outlets <- data.frame(ID = c(31, 3, 5, 1, 45, 92),
                      type = c("outlet", "outlet", "outlet", "terminal", "outlet", "outlet"),
                      stringsAsFactors = FALSE)
aggregated <- aggregate_catchments(walker_fline_rec, walker_catchment_rec, outlets)
plot(aggregated$cat_sets$geom, lwd = 3, border = "red")
plot(walker_catchment_rec$geom, lwd = 1.5, border = "green", col = NA, add = TRUE)
plot(walker_catchment$geom, lwd = 1, add = TRUE)
plot(walker_flowline$geom, lwd = .7, col = "blue", add = TRUE)

plot(aggregated$cat_sets$geom, lwd = 3, border = "black")
plot(aggregated$fline_sets$geom, lwd = 3, col = "red", add = TRUE)
plot(walker_flowline$geom, lwd = .7, col = "blue", add = TRUE)

}
