% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_nhdplusTools.R
\name{align_nhdplus_names}
\alias{align_nhdplus_names}
\title{Align NHD Dataset Names}
\usage{
align_nhdplus_names(x)
}
\arguments{
\item{x}{a \code{sf} object of nhdplus flowlines}
}
\value{
data.frame renamed \code{sf} object
}
\description{
this function takes any NHDPlus dataset and aligns the attribute names with those used in nhdplusTools.
}
\examples{
source(system.file("extdata/new_hope_data.R", package = "nhdplusTools"))

names(new_hope_flowline)

names(new_hope_flowline) <- tolower(names(new_hope_flowline))

new_hope_flowline <- align_nhdplus_names(new_hope_flowline)

names(new_hope_flowline)

}
