% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplusTools.R
\name{align_nhdplus_names}
\alias{align_nhdplus_names}
\title{Align NHD Dataset Names}
\usage{
align_nhdplus_names(x)
}
\arguments{
\item{x}{a \code{sf} object of nhdplus flowlines}
}
\value{
a renamed \code{sf} object
}
\description{
this function takes any NHDPlus dataset and aligns the attribute names with those used in nhdplusTools.
}
\examples{
\dontrun{
a = AOI::getAOI(list("UCSB", 1, 1))
n = HydroData::findNHD(a)[[2]] \%>\% align_nhdplus_names()
UM_comids = get_UM(n, n$COMID[3])
}
}
