% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_network.R
\name{calculate_arbolate_sum}
\alias{calculate_arbolate_sum}
\title{DEPRECATED: Calculate Arbolate Sum}
\usage{
calculate_arbolate_sum(x)
}
\arguments{
\item{x}{data.frame with ID, toID, and length columns.}
}
\value{
numeric with arbolate sum.
}
\description{
Calculates arbolate sum given a dendritic
network and incremental lengths. Arbolate sum is the total length
of all upstream flowlines.
}
