% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_network.R
\name{calculate_total_drainage_area}
\alias{calculate_total_drainage_area}
\title{DEPRECATED: Total Drainage Area}
\usage{
calculate_total_drainage_area(x)
}
\arguments{
\item{x}{data.frame with ID, toID, and area columns.}
}
\value{
numeric with total area.
}
\description{
Calculates total drainage area given a dendritic
network and incremental areas.
}
