% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_catchments.R
\name{collapse_catchments}
\alias{collapse_catchments}
\title{Collapse Catchment Boundaries}
\usage{
collapse_catchments(fline_rec, cat_rec, outlet_cats)
}
\description{
Collapses catchments according to a set of input outlet catchments.
}
\examples{

}
