% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_flowlines}
\alias{collapse_flowlines}
\title{Collapse NHDPlus Network}
\usage{
collapse_flowlines(flines, thresh, add_category = FALSE,
  mainstem_thresh = NULL, exclude_cats = NULL, warn = TRUE)
}
\arguments{
\item{flines}{data.frame with COMID, toCOMID, LENGTHKM, Hydroseq, and LevelPathI columns}

\item{thresh}{numeric a length threshold (km). Flowlines shorter than this
will be collapsed with up or downstream flowlines.}

\item{add_category}{boolean if combination category is desired in output, set
to TRUE}

\item{mainstem_thresh}{numeric threshold for combining inter-confluence
mainstems}

\item{exclude_cats}{integer vector of COMIDs to be excluded from collapse modifications.}

\item{warn}{boolean controls whether warning an status messages are printed}
}
\value{
A refactored network with merged up and down flowlines.
}
\description{
Refactors the NHDPlus flowline network, eliminating short and
non-confluence flowlines. The aim of this function is to create flowpaths
that describe a network of catchments that combines complex hydrology near
confluences into upstream catchments and removes very short flowlines along mainstem
flow-paths.
}
\seealso{
The \code{\link{refactor_nhdplus}} function implements a complete
workflow using `collapse_flowlines()`.
}
