% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse_flowlines}
\alias{collapse_flowlines}
\title{Collapse NHDPlus Network}
\usage{
collapse_flowlines(flines, thresh, add_category = FALSE,
  mainstem_thresh = NULL)
}
\arguments{
\item{flines}{data.frame with COMID, toCOMID, LENGTHKM, and TotDASqKM columns}

\item{thresh}{numeric a length threshold (km). Flowlines shorter than this will be eliminated}

\item{add_category}{boolean if combination category is desired in output, set to TRUE}

\item{mainstem_thresh}{numeric threshold for combining inter-confluence mainstems}
}
\value{
A refactored network with merged up and down flowlines.
}
\description{
Refactors the NHDPlus flowline network, eliminating short and non-cconfluence flowlines.
}
