% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nldi.R
\name{discover_nldi_characteristics}
\alias{discover_nldi_characteristics}
\title{(DEPRECATED) Discover Characteristics Metadata}
\usage{
discover_nldi_characteristics(type = "all")
}
\arguments{
\item{type}{character "all", "local", "total", or "divergence_routed".}
}
\value{
data.frame containing available characteristics
}
\description{
This functionality is deprecated and will be removed in the near future.

Please use \link{get_characteristics_metadata} instead.
}
