% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhdplus.R
\name{download_nhdplushr}
\alias{download_nhdplushr}
\title{Download NHDPlus HiRes}
\usage{
download_nhdplushr(nhd_dir, hu_list, download_files = TRUE)
}
\arguments{
\item{nhd_dir}{character directory to save output into}

\item{hu_list}{character vector of hydrologic region(s) to download}

\item{download_files}{boolean if FALSE, only URLs to files will be returned
can be hu02s and/or hu04s}
}
\value{
Paths to geodatabases created.
}
\description{
Download NHDPlus HiRes
}
\examples{
download_nhdplushr(tempdir(), c("01", "0203"), download_files = FALSE)
}
