% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading_tools.R
\name{download_rf1}
\alias{download_rf1}
\title{Download the seamless Reach File (RF1) Database}
\usage{
download_rf1(
  outdir,
  url = "https://water.usgs.gov/GIS/dsdl/erf1_2.e00.gz",
  progress = TRUE
)
}
\arguments{
\item{outdir}{The folder path where data should be downloaded and extracted}

\item{url}{the location of the online resource}

\item{progress}{boolean display download progress?}
}
\value{
character path to the local e00 file
}
\description{
This function downloads and decompresses staged RF1 data.
See: https://water.usgs.gov/GIS/metadata/usgswrd/XML/erf1_2.xml for metadata.
}
\examples{
\dontrun{
  download_wbd("./data/rf1/")
}
}
