% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloading_tools.R
\name{downloader}
\alias{downloader}
\title{Function to download data from URL to out directory using httr.}
\usage{
downloader(dir, url, type)
}
\arguments{
\item{dir}{path to output directory}

\item{url}{the location of the online resource}

\item{type}{the type of data being downloaded}
}
\value{
the downloaded file path
}
\description{
General downloader
}
