% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaa.R
\name{get_catchment_characteristics}
\alias{get_catchment_characteristics}
\title{Get Catchment Characteristics}
\usage{
get_catchment_characteristics(varname, ids, reference_fabric = "nhdplusv2")
}
\arguments{
\item{varname}{character vector of desired variables}

\item{ids}{numeric vector of identifiers (comids) from the specified fabric}

\item{reference_fabric}{(not used) will be used to allow future specification
of alternate reference fabrics}
}
\description{
Downloads (subsets of) catchment characteristics from a cloud data
store. See \link{get_characteristics_metadata} for available characteristics.

Source:
Wieczorek, M.E., Jackson, S.E., and Schwarz, G.E., 2018, Select Attributes
for NHDPlus Version 2.1 Reach Catchments and Modified Network Routed Upstream
Watersheds for the Conterminous United States (ver. 3.0, January 2021): U.S.
Geological Survey data release, \doi{10.5066/F7765D7V}.
}
