% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{get_DD}
\alias{get_DD}
\title{Get all downstream COMIDs including diversions}
\usage{
get_DD(network, comid, distance = NULL)
}
\arguments{
\item{network}{data.frame NHDPlus flowlines including at a minimum:
COMID, DnMinorHyd, DnHydroseq, and HydroSeq.}

\item{comid}{integer identifier to start navigating from.}

\item{distance}{numeric distance in km to limit how many
COMIDs are returned.
The COMID that exceeds the distance specified is returned.
The longest of the diverted paths is used for limiting distance.}
}
\value{
integer vector of all COMIDs downstream of the starting catchment
}
\description{
Traverse NHDPlus network downstream with diversions
}
\examples{
library(sf)
start_COMID <- 11688818
DD_COMIDs <- get_DD(sample_flines, start_COMID, distance = 4)
plot(dplyr::filter(sample_flines, COMID \%in\% DD_COMIDs)$geometry,
     col = "red", lwd = 2)

DM_COMIDs <- get_DM(sample_flines, start_COMID, distance = 4)
plot(dplyr::filter(sample_flines, COMID \%in\% DM_COMIDs)$geometry,
     col = "blue", add = TRUE, lwd = 2)

}
