% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{get_DM}
\alias{get_DM}
\title{Get all downstream mainstem COMIDs}
\usage{
get_DM(network, comid)
}
\arguments{
\item{network}{data.frame NHDPlus flowlines including at a minimum: COMID,
LevelPathI, DnHydroseq, and HydroSeq.}

\item{comid}{integer identifier to start navigating from.}
}
\value{
integer vector of all COMIDs downstream of the starting catchment along the mainstem.
}
\description{
Traverse NHDPlus network downstream main stem
}
