% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{get_DM}
\alias{get_DM}
\title{Get all downstream mainstem COMIDs}
\usage{
get_DM(network, comid, distance = NULL)
}
\arguments{
\item{network}{data.frame NHDPlus flowlines including at a minimum: COMID,
LevelPathI, DnHydroseq, and HydroSeq.}

\item{comid}{integer identifier to start navigating from.}

\item{distance}{numeric distance in km to limit how many COMIDs are
returned. The COMID that exceeds the distance specified is returned.}
}
\value{
integer vector of all COMIDs downstream of the starting catchment
along the mainstem.
}
\description{
Traverse NHDPlus network downstream main stem
}
\examples{
library(sf)
plot(sample_flines$geometry)
start_COMID <- 11690092
DM_COMIDs <- get_DM(sample_flines, start_COMID)
plot(dplyr::filter(sample_flines, COMID \%in\% DM_COMIDs)$geometry,
     col = "red", add = TRUE, lwd = 3)

DM_COMIDs <- get_DM(sample_flines, start_COMID, distance = 40)
plot(dplyr::filter(sample_flines, COMID \%in\% DM_COMIDs)$geometry,
     col = "blue", add = TRUE, lwd = 2)


}
