% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{get_flowline_index}
\alias{get_flowline_index}
\title{get flowline index}
\usage{
get_flowline_index(flines, points, search_radius = 0.1)
}
\arguments{
\item{flines}{sf data.frame of type LINESTRING or MULTILINESTRING including
COMID, reachcode, ToMeas, and FromMeas}

\item{points}{sfc of type POINT}

\item{search_radius}{numeric the distance for the nearest neighbor search to extend.
See RANN nn2 documentation for more details.}
}
\value{
data.frame with three columns, COMID, reachcode, and REACH_meas.
}
\description{
given an sf point geometry collumn, return COMID, reachcode, and measure for each.
}
\details{
Note 1: that inputs are cast into LINESTRINGS. Because of this, the measure output
of inputs that are true multipart lines may be in error.

Note 2: This version of this function only resolves to the nearest geometry node. This can
cause problems if there are not enough nodes on a given flowline.
}
