% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geoconnex.R
\name{get_geoconnex_reference}
\alias{get_geoconnex_reference}
\title{get geoconnex reference feature layers}
\usage{
get_geoconnex_reference(
  AOI = NULL,
  ids = NULL,
  type = NULL,
  t_srs = NULL,
  buffer = 0.5,
  status = TRUE
)
}
\arguments{
\item{AOI}{sf (MULTI)POINT or (MULTI)POLYGON. An 'area of interest' can
be provided as either a location (sf POINT) or area (sf POLYGON)
in any Spatial Reference System.}

\item{ids}{character or numeric. A set of identifier(s) from the data
type requested, for example if NHD, then a set of COMID(s).}

\item{type}{character. Type of feature to return
('huc08','huc12', 'nhd', 'catchment', 'waterbodies', 'gagesII').
If NULL (default) a data.frame of available resources is returned}

\item{t_srs}{character (PROJ string or EPSG code) or numeric (EPSG code).
A user specified - target -Spatial Reference System (SRS/CRS) for returned objects.
Will default to the CRS of the input AOI if provided, and to 4326 for ID requests.}

\item{buffer}{numeric. The amount (in meters) to buffer a POINT AOI by for an
extended search. Default = 0.5}

\item{status}{boolean print status or not}
}
\value{
sf data.frame containing requested reference features
}
\description{
Queries the geoconnex reference feature server for features of interest.
}
\examples{
\donttest{

dplyr::distinct(discover_geoconnex_reference()[c("id", "title")])

AOI <- sf::st_as_sfc(sf::st_bbox(c(xmin = -89.56684, ymin = 42.99816,
                                   xmax = -89.24681, ymax = 43.17192),
                                 crs = "+proj=longlat +datum=WGS84 +no_defs"))

get_geoconnex_reference(AOI, type = "hu04")

get_geoconnex_reference(type = "hu04", ids = c("0707", "0709"))

AOI <- sf::st_sfc(sf::st_point(c(-89.56684, 42.99816)),
                  crs = "+proj=longlat +datum=WGS84 +no_defs")

get_geoconnex_reference(AOI, type = "hu04", buffer = 100000)

}
}
