% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hydro.R
\name{get_huc12}
\alias{get_huc12}
\title{Find WBD HUC 12 unit subsets}
\usage{
get_huc12(AOI = NULL, id = NULL)
}
\arguments{
\item{id}{WBD HUC12 ID(s)}
}
\value{
simple feature polygon object
}
\description{
Subsets the WBD level 12 features by location (POINT),
area (POLYGON), or set of IDs.
}
