% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainstems.R
\name{get_mainstem_lookup_path}
\alias{get_mainstem_lookup_path}
\title{File path to mainstem lookup table cache}
\usage{
get_mainstem_lookup_path(type)
}
\arguments{
\item{type}{character one of "nhdpv2" or "nhdphr"}
}
\value{
character file path
}
\description{
hydrogeofetch will download and cache a parquet file
translating geoconnex mainstem identifiers to NHDPlusV2 comids
(\code{type = "nhdpv2"}) or NHDPlusHR NHDPlusIDs (\code{type = "nhdphr"}).
Will use the user data dir indicated by \link{hydrogeofetch_data_dir}.
}
\keyword{internal}
