% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hydro.R
\name{get_nhdarea}
\alias{get_nhdarea}
\title{Find NHD Areas}
\usage{
get_nhdarea(AOI = NULL, id = NULL)
}
\arguments{
\item{AOI}{sf (MULTI)POINT or (MULTI)POLYGON. An 'area of interest' can
be given as either a location (sf POINT) or area (sf POLYGON)
in any coordinate reference system.}

\item{id}{Search for NHD Areas Bodies by COMID}
}
\value{
simple feature polygon object
}
\description{
Subsets NHD Area features by location (POINT),
area (POLYGON), or set of IDs.
}
