% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhdplus.R
\name{get_nhdplus}
\alias{get_nhdplus}
\title{Get National Hydrography Subsets (Multirealization)}
\usage{
get_nhdplus(
  AOI = NULL,
  comid = NULL,
  nwis = NULL,
  realization = "flowline",
  streamorder = NULL
)
}
\arguments{
\item{AOI}{sf (MULTI)POINT or (MULTI)POLYGON. An 'area of interest' can
be given as either a location (sf POINT) or area (sf POLYGON)
in any coordinate reference system.}

\item{comid}{numeric or character. Search for NHD features by COMID(s)}

\item{nwis}{numeric or character. Search for NHD features by
colocated NWIS identifiers}

\item{realization}{character. What realization to return.
Default is flowline and options include: outlet, flowline, catchment,
and all}

\item{streamorder}{numeric or character. Only return NHD flowlines with a
streamorder greater then or equal to this value
for input value and higher.
Only usable with AOI and flowline realizations.}
}
\value{
a single, or list, or simple feature objects
}
\description{
Subsets NHD features by location (POINT), area (POLYGON),
or set of COMIDs. Multi realizations are supported allowing you to query
for flowlines, catchments, or outlets.
}
\examples{
 point <- sf::st_sfc(sf::st_point(c(-119.845, 34.4146)), crs = 4326)
 get_nhdplus(point)
 get_nhdplus(point, realization = "catchment")
 get_nhdplus(point, realization = "all")
 get_nhdplus(comid = 101)
 get_nhdplus(nwis  = c(11120000, 11120500))
 area <- sf::st_as_sfc(sf::st_bbox(c(xmin = -119.8851, xmax =-119.8361,
 ymax = 34.42439, ymin = 34.40473), crs = 4326))
 get_nhdplus(area)
 get_nhdplus(area, realization = "flowline", streamorder = 3)
}
