% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhdplus.R
\name{get_nhdplushr}
\alias{get_nhdplushr}
\title{Get NHDPlus HiRes as single geopackage}
\usage{
get_nhdplushr(hr_dir, out_gpkg)
}
\arguments{
\item{hr_dir}{character directory with geodatabases}

\item{out_gpkg}{character path to write output geopackage}
}
\description{
Get NHDPlus HiRes as single geopackage
}
