% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nldi.R
\name{get_nldi_basin}
\alias{get_nldi_basin}
\title{Get NLDI Basin Boundary}
\usage{
get_nldi_basin(nldi_feature, simplify = TRUE, split = FALSE)
}
\arguments{
\item{nldi_feature}{list with names `featureSource` and `featureID` where
`featureSource` is derived from the "source" column of  the response of
\link[dataRetrieval]{get_nldi_sources} and the `featureID` is a known identifier
from the specified `featureSource`.}

\item{simplify}{logical should response geometry be simplified for
visualization and performance?}

\item{split}{logical should response resolve precisely to the location
of the `nldi_feature`? Setting `TRUE` calls an additional service and
will be slower and less robust.}
}
\value{
sf data.frame with result basin boundary
}
\description{
Get a basin boundary for a given NLDI feature.
}
\details{
Only resolves to the nearest NHDPlus catchment divide. See:
https://waterdata.usgs.gov/blog/nldi-intro/ for more info on the nldi.
}
