% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_pathlength}
\alias{get_pathlength}
\title{Get path length}
\usage{
get_pathlength(x)
}
\arguments{
\item{x}{data.frame with ID, toID, length columns.}
}
\value{
data.frame containing levelpaths for each ID
}
\description{
Generates the main path length to a basin's
terminal path.
}
\examples{
source(system.file("extdata", "walker_data.R", package = "nhdplusTools"))

fl <- dplyr::select(prepare_nhdplus(walker_flowline, 0, 0),
                    ID = COMID, toID = toCOMID, length = LENGTHKM)

get_pathlength(fl)

}
