% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_network.R
\name{get_pfaf}
\alias{get_pfaf}
\title{Get Pfafstetter Codes (Experimental)}
\usage{
get_pfaf(fl, max_level = 2)
}
\arguments{
\item{fl}{sf data.frame with ID, toID, totda, outletID, topo_sort,
and levelpath attributes.}

\item{max_level}{integer number of pfaf levels to attempt to calculate.
If the network doesn't have resolution to support the desired level,
unexpected behavior may occur.}
}
\value{
data.frame with ID and pfaf columns.
}
\description{
Determines Pfafstetter codes for a dendritic network with
total drainage area, levelpath, and topo_sort attributes.
}
\examples{
library(dplyr)
source(system.file("extdata/nhdplushr_data.R", package = "nhdplusTools"))
hr_flowline <- nhdplusTools:::rename_nhdplus(hr_flowline)

fl <- prepare_nhdplus(hr_flowline, 0, 0, purge_non_dendritic = TRUE, warn = FALSE) \%>\%
  left_join(select(hr_flowline, COMID, AreaSqKM), by = "COMID") \%>\%
  sf::st_sf() \%>\%
  select(ID = COMID, toID = toCOMID, area = AreaSqKM)

fl$nameID = ""
fl$totda <- calculate_total_drainage_area(sf::st_set_geometry(fl, NULL))
fl <- left_join(fl, get_levelpaths(rename(sf::st_set_geometry(fl, NULL), weight = totda)), by = "ID")

pfaf <- get_pfaf(fl, max_level = 2)

fl <- left_join(fl, pfaf, by = c("ID" = "members"))

plot(fl["pf_level_2"], lwd = 2)

source(system.file("extdata", "walker_data.R", package = "nhdplusTools"))

fl <- prepare_nhdplus(walker_flowline, 0, 0, purge_non_dendritic = FALSE, warn = FALSE) \%>\%
  left_join(select(walker_flowline, COMID, AreaSqKM), by = "COMID") \%>\%
  sf::st_sf() \%>\%
  select(ID = COMID, toID = toCOMID, area = AreaSqKM)

fl$nameID = ""
fl$totda <- calculate_total_drainage_area(sf::st_set_geometry(fl, NULL))
fl <- left_join(fl, get_levelpaths(rename(sf::st_set_geometry(fl, NULL), weight = totda)), by = "ID")

pfaf <- get_pfaf(fl, max_level = 2)

fl <- left_join(fl, pfaf, by = c("ID" = "members"))

plot(fl["pf_level_2"], lwd = 2)

}
