% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oaproc.R
\name{get_raindrop_trace}
\alias{get_raindrop_trace}
\title{Get Raindrop Trace}
\usage{
get_raindrop_trace(point, direction = "down")
}
\arguments{
\item{point}{sfc POINT including crs as created by:
\code{sf::st_sfc(sf::st_point(.. ,..), crs)}}

\item{direction}{character \code{"up"}, \code{"down"}, or \code{"none"}.
Controls the portion of the split flowline that is returned along with
the raindrop trace line.}
}
\value{
sf data.frame containing raindrop trace and requested
portion of flowline.
}
\description{
Uses a raindrop trace web service to trace the
nhdplus digital elevation model to the nearest downslope flowline.
}
