% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_sorted}
\alias{get_sorted}
\title{Get Sorted Network}
\usage{
get_sorted(x, split = FALSE, outlets = NULL)
}
\arguments{
\item{x}{data.frame with an identifier and to identifier in the
first and second columns.}

\item{split}{logical if TRUE, the result will be split into
independent networks identified by the id of their outlet. The
outlet id of each independent network is added as a "terminalID"
attribute.}

\item{outlets}{same as id in x; if specified only the network
eminating from these outlets will be considered and returned.}
}
\value{
data.frame containing a topologically sorted version
of the requested network and optionally a terminal id.
}
\description{
given a tree with an id and and toid in the
first and second columns, returns a sorted and potentially
split set of output.
}
