% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_network.R
\name{get_terminal}
\alias{get_terminal}
\title{Get Terminal ID}
\usage{
get_terminal(x, outlets)
}
\arguments{
\item{x}{two column data.frame with IDs and toIDs. Names are ignored.}

\item{outlets}{IDs of outlet flowlines}
}
\description{
Get the ID of the basin outlet for each flowline.
}
