% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network.R
\name{get_UM}
\alias{get_UM}
\title{Get all upstream mainstem COMIDs}
\usage{
get_UM(network, comid, distance = NULL)
}
\arguments{
\item{network}{data.frame NHDPlus flowlines including at a minimum: COMID,
Pathlength, LevelPathI, UsHydroseq, and HydroSeq.}

\item{comid}{integer identifier to start navigating from.}

\item{distance}{numeric distance in km to limit how many COMIDs are returned. The COMID that exceeds the distance specified is returned.}
}
\value{
integer vector of all COMIDs upstream of the starting catchment along the mainstem.
}
\description{
Traverse NHDPlus network upstream main stem
}
