% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaa.R
\name{get_vaa}
\alias{get_vaa}
\title{NHDPlusV2 Attribute Subset}
\usage{
get_vaa(atts = NULL)
}
\arguments{
\item{atts}{The variable names you would like, always includes comid}
}
\value{
data.frame
}
\description{
Return requested NHDPlusv2 Attributes
}
\examples{
get_vaa("slope")
get_vaa(c("slope", "lengthkm"))
}
