% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaa.R
\name{get_vaa}
\alias{get_vaa}
\title{NHDPlusV2 Attribute Subset}
\usage{
get_vaa(atts = NULL, path = get_vaa_path(), download = TRUE)
}
\arguments{
\item{atts}{character The variable names you would like, always includes comid}

\item{path}{character path where the file should be saved. Default is a
persistent system cache as retrieved by \link[rappdirs]{user_cache_dir}.
Also see: \link{get_vaa_path}}

\item{download}{logical if TRUE, the default, will download VAA table if not
found at path.}
}
\value{
data.frame
}
\description{
Return requested NHDPlusv2 Attributes
}
\examples{
\dontrun{
# NOTE: path = tempfile() for demo only.
temp <- tempfile()
get_vaa("slope", path = temp)
get_vaa(c("slope", "lengthkm"), path = temp)
}
}
