% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vaa.R
\name{get_vaa_path}
\alias{get_vaa_path}
\title{File path to value added attribute (vaa) Cache}
\usage{
get_vaa_path(updated_network = FALSE)
}
\arguments{
\item{updated_network}{logical default FALSE. If TRUE, returns path to updated
network parameters. See \link{get_vaa} for more.}
}
\value{
character file path
}
\description{
nhdplusTools will download and cache an `fst` file with
NHDPlusV2 attribute data sans geometry. This function returns the
file path to the cached file. Will use the user data dir indicated
by \link{nhdplusTools_data_dir}.
}
\details{
The VAA data is a aggregate table of information from the NHDPlusV2
elevslope.dbf(s), PlusFlowlineVAA.dbf(s); and NHDFlowlines. All data
originates from the EPA NHDPlus Homepage
\href{https://www.epa.gov/waterdata/get-nhdplus-national-hydrography-dataset-plus-data}{here}.
To see the location of cached data on your machine use
\code{\link{get_vaa_path}}.
To view aggregate data and documentation, see
\href{https://www.hydroshare.org/resource/6092c8a62fac45be97a09bfd0b0bf726/}{here}
}
\examples{
get_vaa_path()

get_vaa_path(updated_network = TRUE)

}
