% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_watershed_delineation_nhdplushr.R
\name{get_watershed_by_huc}
\alias{get_watershed_by_huc}
\title{get_watershed_by_huc}
\usage{
get_watershed_by_huc(points_same_huc, nhdplusdir)
}
\arguments{
\item{points_same_huc}{an sf object of type POINT or MULTIPOINT containing the the locations
of desired catchment outlets. These should all be from the same huc4 (scale of
nhdplushr data download)}

\item{nhdplusdir}{directory location of nhdplushr data downloads}
}
\value{
catchments for all the points within a huc4
}
\description{
function that takes input of points all in the same HUC4, downloads the
HUC4 flowlines and catchment data from nhdplushr, and outputs catchments.
}
