% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oaproc.R
\name{get_xs_points}
\alias{get_xs_points}
\title{Get Cross Section Endpoints (experimental)}
\usage{
get_xs_points(point1, point2, num_pts, res = 1)
}
\arguments{
\item{point1}{sfc POINT including crs as created by:
\code{sf::st_sfc(sf::st_point(.. ,..), crs)}}

\item{point2}{sfc POINT including crs.}

\item{num_pts}{numeric number of points to retrieve along the cross section.}

\item{res}{integer resolution of 3D Elevation Program data to request.
Must be on of: 1, 3, 5, 10, 30, 60.}
}
\value{
sf data.frame containing points retrieved.
}
\description{
Uses a cross section retrieval web services to retrieve a
cross section between two endpoints.
}
