% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_topology.R
\name{make_node_topology}
\alias{make_node_topology}
\title{get node topology from edge topology}
\usage{
make_node_topology(x, add = TRUE)
}
\arguments{
\item{x}{data.frame with an identifier and to identifier in the
first and second columns.}

\item{add}{logical if TRUE, a additional data will be added, otherwise
a data.frame with only an identifier and new attributes wil be returned.}
}
\value{
data.frame containing id, fromnode, and tonode attributes or all
attributes provided with id, fromnode and tonodde in the first three columns.
}
\description{
creates a node topology table from an edge topology
}
\examples{
source(system.file("extdata/new_hope_data.R", package = "nhdplusTools"))

x <- get_tocomid(
  dplyr::select(new_hope_flowline, COMID, FromNode, ToNode, Divergence, FTYPE,
                AreaSqKM, LENGTHKM, GNIS_ID)
)

head(make_node_topology(x))
}
