% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhdplushr.R
\name{make_standalone}
\alias{make_standalone}
\title{Make isolated NHDPlusHR region a standalone dataset}
\usage{
make_standalone(flowlines, fix_terminals = TRUE)
}
\arguments{
\item{flowlines}{sf data.frame of NHDPlusHR flowlines.}

\item{fix_terminals}{boolean If true, an attempt is made to make flowlines that
leave the domain act as terminal paths.}
}
\description{
Cleans up and prepares NHDPlusHR regional data for use as complete NHDPlus
data. The primary modification applied is to ensure that any flowpath that exits the
domain is labeled as a terminal path and attributes are propogated upstream such that
the domain is independently complete.
}
\examples{
source(system.file("extdata/nhdplushr_data.R", package = "nhdplusTools"))

(outlet <- hr_flowline[hr_flowline$Hydroseq == min(hr_flowline$Hydroseq), ])
length(sf::st_geometry(hr_flowline)[hr_flowline$TerminalPa == outlet$Hydroseq])

hr_flowline <- make_standalone(hr_flowline)

(outlet <- hr_flowline[hr_flowline$Hydroseq == min(hr_flowline$Hydroseq), ])
length(sf::st_geometry(hr_flowline)[hr_flowline$TerminalPa == outlet$Hydroseq])



}
