% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nhdplushr.R
\name{make_standalone}
\alias{make_standalone}
\title{Make isolated NHDPlusHR region a standalone dataset}
\usage{
make_standalone(flowlines, fix_terminals = TRUE)
}
\arguments{
\item{flowlines}{sf data.frame of NHDPlusHR flowlines.}

\item{fix_terminals}{boolean If true, an attempt is made to make flowlines that
leave the domain act as terminal paths.}
}
\description{
Cleans up and prepares NHDPlusHR regional data for use as complete NHDPlus
data. The primary modification applied is to ensure that any flowpath that exits the
domain is labeled as a terminal path and attributes are propogated upstream such that
the domain is independently complete.
}
