% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_levelpaths.R
\name{match_levelpaths}
\alias{match_levelpaths}
\title{Match Level Paths}
\usage{
match_levelpaths(fline_hu, start_comid)
}
\arguments{
\item{fline_hu}{sf data.frame flowlines intersected with hydrologic units}

\item{start_comid}{integer COMID to start search from.}
}
\description{
Attempts to match dendritic hydrologic unit networks to NHDPlus
flowline level paths.
}
\details{
...
}
\examples{
#todo

}
