% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_levelpaths.R
\name{match_levelpaths}
\alias{match_levelpaths}
\title{Match Level Paths}
\usage{
match_levelpaths(fline_hu, start_comid, add_checks = FALSE)
}
\arguments{
\item{fline_hu}{sf data.frame flowlines intersected with hydrologic units containing
COMID, Hydroseq, LevelPathI, DnLevelPat, denTotalAreaSqKM, HUC12, TOHUC attributes.}

\item{start_comid}{integer COMID to start search from.}

\item{add_checks}{boolean if TRUE, checks for toHUC errors are added.}
}
\description{
Attempts to match dendritic hydrologic unit networks to level paths.
}
\details{
...
}
\examples{
#todo

}
