% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhdplus_refactor.R
\name{prepare_nhdplus}
\alias{prepare_nhdplus}
\title{Prep NHDPlus Data for Refactor}
\usage{
prepare_nhdplus(flines, min_network_size, min_path_length,
  purge_non_dendritic = TRUE)
}
\arguments{
\item{flines}{data.frame NHDPlus flowlines including at a minimum: COMID, LENGTHKM,
FTYPE, TerminalFl, FromNode, ToNode, TotDASqKM, StartFlag, StreamOrde,
StreamCalc, TerminalPa, and Pathlength variables.}

\item{min_network_size}{numeric Minimum size (sqkm) of drainage network to include in output.}

\item{min_path_length}{numeric Minimum length (km) of terminal level path of a network.}

\item{purge_non_dendritic}{boolean Should non dendritic paths be removed or not.}
}
\value{
data.frame ready to be used with the refactor_flowlines function.
}
\description{
Function to prep NHDPlus data for network refactoring
}
