% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconcile.R
\name{reconcile_collapsed_flowlines}
\alias{reconcile_collapsed_flowlines}
\title{Reconcile Collapsed Flowlines}
\usage{
reconcile_collapsed_flowlines(flines, geom = NULL, id = "COMID")
}
\arguments{
\item{flines}{data.frame with COMID, toCOMID, LENGTHKM, LevelPathI, Hydroseq,
and TotDASqKM columns}

\item{geom}{sf data.frame for flines}

\item{id}{character id collumn name.}
}
\value{
reconciled flowpaths with new ID, toID, LevelPathID, and Hydroseq identifiers.
Note that all the identifiers are new integer IDs. LevelPathID and Hydroseq are consistent
with the LevelPathID and Hydroseq from the input NHDPlus flowlines.
}
\description{
Reconciles output of collapse_flowlines giving a unique ID to
each new flowpath and providing a mapping to NHDPlus COMIDs.
}
\seealso{
The \code{\link{refactor_nhdplus}} function implements a complete
workflow using `reconcile_collapsed_flowlines()`.
}
