% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor_nhdplus.R
\name{refactor_nhdplus}
\alias{refactor_nhdplus}
\title{Refactor NHDPlus}
\usage{
refactor_nhdplus(nhdplus_flines, split_flines_meters, split_flines_cores,
  collapse_flines_meters, collapse_flines_main_meters, out_collapsed,
  out_reconciled, three_pass = FALSE, purge_non_dendritic = TRUE,
  warn = TRUE)
}
\arguments{
\item{nhdplus_flines}{data.frame raw nhdplus flowline features as
derived from the national seamless geodatabase.}

\item{split_flines_meters}{numeric the maximum length flowline desired
in the output.}

\item{split_flines_cores}{numeric the number of processing cores to use
while splitting flowlines.}

\item{collapse_flines_meters}{numeric the minimum length of
inter-confluence flowline desired in the output.}

\item{collapse_flines_main_meters}{numeric the minimum length of
between-confluence flowlines.}

\item{out_collapsed}{character where to write a geopackage containing
the split and collapsed flowlines.}

\item{out_reconciled}{character where to write a geopackage containing
the reconciled flowlines.}

\item{three_pass}{boolean whether to perform a three pass collapse or
single pass.}

\item{purge_non_dendritic}{boolean passed on to prepare_nhdplus}

\item{warn}{boolean controls whether warning an status messages are printed}
}
\description{
A complete network refactor workflow has been packaged
into this function. See details and vignettes for more information.
}
\details{
This is a convenient wrapper function that implements three phases
of the network refactor workflow: split, collapse, reconcile. See the
NHDPlus Refactor vignette for details of these three steps by running:
\code{vignette("refactor_nhdplus", package = "nhdplusTools")}
}
\examples{
nhdplus_flowlines <- sf::st_zm(sample_flines)
refactor_nhdplus(nhdplus_flines = nhdplus_flowlines,
                 split_flines_meters = 2000,
                 split_flines_cores = 3,
                 collapse_flines_meters = 500,
                 collapse_flines_main_meters = 500,
                 out_collapsed = "temp.gpkg",
                 out_reconciled = "temp_rec.gpkg",
                 three_pass = TRUE,
                 purge_non_dendritic = FALSE,
                 warn = FALSE)
unlink("temp.gpkg")
unlink("temp_rec.gpkg")

}
\seealso{
The following four functions are used in the `refactor_nhdplus` workflow.
\enumerate{
  \item \code{\link{prepare_nhdplus}}
  \item \code{\link{split_flowlines}}
  \item \code{\link{collapse_flowlines}}
  \item \code{\link{reconcile_collapsed_flowlines}}
}
}
