% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_flowline.R
\name{split_flowlines}
\alias{split_flowlines}
\title{Split Flowlines}
\usage{
split_flowlines(flines, max_length, para = 0)
}
\arguments{
\item{flines}{data.frame with COMID, toCOMID, LENGTHKM
and LINESTRING sf column in "meters" projection}

\item{max_length}{maximum segment length to return}

\item{para}{numeric how many threads to use in parallel computation}
}
\value{
All the flowlines with some split apart.
}
\description{
A wrapper for split_lines that works on nhdplus attributes
}
\seealso{
The \code{\link{refactor_nhdplus}} function implements a complete
workflow using `split_flowlines()`.
}
