% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_lines}
\alias{split_lines}
\title{split lines}
\usage{
split_lines(input_lines, max_length, id = "ID", para = 0)
}
\arguments{
\item{input_lines}{data.frame of class sf with LINESTRING sfc column.}

\item{max_length}{maximum segment length to return}

\item{id}{name of ID column in data.frame}

\item{para}{how many cores to use}
}
\value{
only the split lines.
}
\description{
Splits lines longer than a given threshold into the minimum number of pieces to all be under the given threshold.
}
