% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_nhdplus}
\alias{subset_nhdplus}
\title{Subset NHDPlus National Data}
\usage{
subset_nhdplus(comids, output_file, nhdplus_data = NULL,
  simplified = TRUE, overwrite = FALSE, status = TRUE)
}
\arguments{
\item{comids}{integer vector of COMIDs to include.}

\item{output_file}{character path to save the output to defaults
to the directory of the nhdplus_data.}

\item{nhdplus_data}{character path to the .gpkg or .gdb containing
the national seamless dataset or "download" to use web service.
Not required if \code{\link{nhdplus_path}} has been set or the default
has been adopted. See details for more.}

\item{simplified}{boolean if TRUE (the default) the CatchmentSP layer
will be included. Not relevant to the "download" option.}

\item{overwrite}{boolean should the output file be overwritten}

\item{status}{boolean should the function print status messages}
}
\value{
path to the saved subset geopackage
}
\description{
Saves a subset of the National Seamless database based on a
specified collection of catchments.
}
\details{
If \code{\link{stage_national_data}} has been run in the current
session, this function will use the staged national data automatically.

This function relies on the National Seamless Geodatabase or Geopackage.
It can be downloaded
\href{https://www.epa.gov/waterdata/nhdplus-national-data}{here.}

The "download" option of this function should be considered preliminary
and subject to revision. It does not include as many layers and may not
be available permenantly.
}
\examples{
sample_data <- system.file("extdata/sample_natseamless.gpkg",
                           package = "nhdplusTools")

nhdplus_path(sample_data)

staged_nhdplus <- stage_national_data(output_path = "./")

sample_flines <- readRDS(staged_nhdplus$flowline)

geom_col <- attr(sample_flines, "sf_column")

plot(sample_flines[[geom_col]],
     lwd = 3)

start_point <- sf::st_sfc(sf::st_point(c(-89.362239, 43.090266)),
                          crs = 4326)

plot(start_point, cex = 1.5, lwd = 2, col = "red", add = TRUE)

start_comid <- discover_nhdplus_id(start_point)

comids <- get_UT(sample_flines, start_comid)

plot(dplyr::filter(sample_flines, COMID \%in\% comids)[[geom_col]],
     add=TRUE, col = "red", lwd = 2)

output_file <- "./demo_subset.gpkg"

subset_nhdplus(comids = comids,
               output_file = output_file,
               nhdplus_data = sample_data,
               overwrite = TRUE,
               status = TRUE)

sf::st_layers(output_file)

catchment <- sf::read_sf(output_file, "CatchmentSP")

plot(catchment[[attr(catchment, "sf_column")]], add = TRUE)

waterbody <- sf::read_sf(output_file, "NHDWaterbody")

plot(waterbody[[attr(waterbody, "sf_column")]],
     col = rgb(0, 0, 1, alpha = 0.5), add = TRUE)

# Download Option:
subset_nhdplus(comids = comids,
               output_file = output_file,
               nhdplus_data = "download",
               overwrite = TRUE,
               status = TRUE)

sf::st_layers(output_file)

}
